requirements.txt
atlassian-python-api==3.14.1
certifi==2021.10.8
charset-normalizer==2.0.9
defusedxml==0.7.1
Deprecated==1.2.13
idna==3.3
importlib-metadata==4.8.2
keyring==23.4.0
oauthlib==3.1.1
pywin32-ctypes==0.2.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
urllib3==1.26.7
wrapt==1.13.3
zipp==3.6.0
